refactor(kirra-map): add the split submodule files + mod.rs content
The prior commit (328a1de) recorded only the lanemap.rs -> lanemap/mod.rs rename
with stale content; a bad pathspec had aborted `git add`, so the actual split
(mod.rs body + occlusion.rs/junctions.rs/routing.rs) was left unstaged. This
commit lands the real content — the working tree that built + tested green.
